statistics crawling support and views

+++ stats/management/commands/fetch_standings.py
@@ -7,7 +7,7 @@ from django.core.management import BaseCommand
 
 from stats.models import StandingsSet, StandingsRow, StatsTeam
 
-INDIANS_ID = 2923
+SERIES_ID = 2923
 
 
 class StandingsHTMLParser(HTMLParser):
@@ -21,7 +21,7 @@ class StandingsHTMLParser(HTMLParser):
     __digit_counter = None
 
     def handle_starttag(self, tag, attrs):
-        print("Encountered a start tag:", tag)
+        # print("Encountered a start tag:", tag)
         self.__cur_elem = None
         if tag == 'td':
             if ('class', 'ranking') in attrs:
@@ -37,12 +37,12 @@ class StandingsHTMLParser(HTMLParser):
                 self.__cur_elem = 'digit'
 
     def handle_endtag(self, tag):
-        print("Encountered an end tag :", tag)
+        # print("Encountered an end tag :", tag)
         if tag == 'body':
             self.team_data.append(self.__cur_team_data)
 
     def handle_data(self, data):
-        print("Encountered some data  :", data)
+        # print("Encountered some data  :", data)
         if self.__cur_elem:
             data = data.strip()
             if data:
@@ -62,27 +62,28 @@ class Command(BaseCommand):
 
     def handle(self, *args, **kwargs):
         dryrun = kwargs['dryrun']
-        url = f"https://api.floorball.fi/embed/{INDIANS_ID}/standings"
+        url = f"https://api.floorball.fi/embed/{SERIES_ID}/standings"
         if url.startswith('http'):
             print('Fetching from URL %s' % url)
             with requests.get(url) as r:
                 parser = StandingsHTMLParser()
                 parser.feed(r.content.decode('UTF-8'))
                 pprint.pprint(parser.team_data)
-                recent_rows = StandingsSet.objects.filter(created_at__gte=(datetime.datetime.now() - datetime.timedelta(hours=6)))
-                for recent in recent_rows:
-                    self.stdout.write(self.style.WARNING(f"Deleting recent set {recent} to create an update"))
-                    recent.delete()
-                standings_set = StandingsSet.objects.create()
-                for team_data in parser.team_data:
-                    team = StatsTeam.get_or_create(list(filter(lambda x: x[0] == 'name', team_data))[0])
-                    row = StandingsRow(set=standings_set, team=team)
-                    for attr, val in team_data:
-                        if hasattr(row, attr):
-                            setattr(row, attr, val)
-                    # note that we are hardcoding this here to 0 as it is not provided in the data
-                    row.wins_contd_period = 0
-                    row.save()
+                if not dryrun:
+                    recent_rows = StandingsSet.objects.filter(created_at__gte=(datetime.datetime.now() - datetime.timedelta(hours=6)))
+                    for recent in recent_rows:
+                        self.stdout.write(self.style.WARNING(f"Deleting recent set {recent} to create an update"))
+                        recent.delete()
+                    standings_set = StandingsSet.objects.create()
+                    for team_data in parser.team_data:
+                        team = StatsTeam.get_or_create(list(filter(lambda x: x[0] == 'name', team_data))[0][1])
+                        row = StandingsRow(set=standings_set, team=team)
+                        for attr, val in team_data:
+                            if hasattr(row, attr):
+                                setattr(row, attr, val)
+                        # note that we are hardcoding this here to 0 as it is not provided in the data
+                        row.wins_contd_period = 0
+                        row.save()
 
-                self.stdout.write(self.style.SUCCESS(f"Created a new Standings set with {standings_set.rows.count()} rows"))
+                    self.stdout.write(self.style.SUCCESS(f"Created a new Standings set with {standings_set.rows.count()} rows"))
